Port MidasWWW 2.1 to 64-bit Linux with Motif 2.3
Builds with gcc 15 and libmotif-dev on Ubuntu 26.04 and runs: renders
HTML over HTTP and from local files, follows links, history, menus,
dialogs, Document Source, scrolling and clean exit all work. Verified
under AddressSanitizer.

LP64 fixes: real return types for every function (midasproto.h), a
pointer-sized integer member in the interpreter operand union and a
matching XtConvertAndStore result buffer, pointer-sized lengths where an
int was smuggled through a char ** (dump_text and the selection dumpers),
correct resource size for SGMLNtagList, correct allocation for the
gethostbyname cache, SGMLText num_popups alias offset on LP64, unsigned
widget hashing, NULL -> 0 for XID/quark/interval values, cuserid and
sys_errlist replacements, and dropping the local BSD getenv/setenv.

Motif 2.x fixes: DynamicPixmap and Set converters, and the
scrolledWindowChildType constraint so XmMainWindow lays out the header
and button bar instead of swallowing them into the clip window.

Other: scan widget paths on a private copy (callers pass string literals
which are read-only now), fix a conditional XtSetArg that left an
uninitialised Arg, fix a one-byte-before-buffer read in the URL parser,
and treat file:/// and file://localhost/ as local files.

Adds a Linux Makefile, PORTING-LINUX.md and a .gitignore; uid2c is a
build product and is no longer tracked.

+++ midaswww-2.1/SGMLFormattedText.c
@@ -975,7 +975,7 @@ char **pp;
    int maxlen = 80;
    int width;
 
-   if (dp->calc_len) { len = (int) *pp; width = len - dp->last_len; }
+   if (dp->calc_len) { len = (int) (long) *pp; width = len - dp->last_len; }
    else              { p = *pp;         width = p - dp->last_pos;   }
 
    switch (dp->format_type)
@@ -1035,7 +1035,7 @@ char **pp;
                if (dp->copy_type == SGMLSELECT_LINE &&
                    w == (SGMLFormattedTextObject) dp->sw) len++;   /* explicitely add a line */
 
-               *pp = (char *)len;
+               *pp = (char *) (long) len;
              }
            else
              {
